use simpler usePatchOrganization, merge files
